Get list edit page working
Needed to fix bug in total cost updating. Now we have to get all the items in the current list and sum the costs of the ones which are bought. Previously we were leaving that to SQL but the LiveData was not updating as expected when the list was changed.